feat(outputs): add Windows DXGI shared texture foundation for Spout output sink
- Add WindowsSurface in PluginSink using D3D12 shared committed resources and DXGI shared handles
- Wrap shared DXGI textures in wgpu::Texture via wgpu-hal dx12 backend
- Expand surface_id to u64 in output plugin wire protocol (HostMessage::Frame) for 64-bit Windows HANDLEs
- Update Outputs bay chips in console view to display Spout on Windows and Syphon on macOS
- Add Spout candidate binary resolution and availability checks in operations
- Add Windows DXGI shared handle allocation unit test in plugin_sink

+++ crates/karakuri-environment/src/output_plugin/message.rs
@@ -29,8 +29,8 @@ pub enum HostMessage {
         index: u64,
         /// Platform-native shareable surface identifier:
         /// - On macOS: `IOSurfaceID` (u32) resolvable via `IOSurfaceLookup`
-        /// - On Windows: DXGI shared handle (u32 or u64)
-        surface_id: u32,
+        /// - On Windows: DXGI shared handle (`HANDLE`, 64-bit integer)
+        surface_id: u64,
         /// Current frame width in pixels.
         width: u32,
         /// Current frame height in pixels.
